Refactor médico
