+++ wbhealth/src/main/java/br/com/dbc/wbhealth/controller/MedicoController.java
@@ -2,6 +2,7 @@ package br.com.dbc.wbhealth.controller;
 
 import br.com.dbc.wbhealth.documentation.MedicoControllerDoc;
 import br.com.dbc.wbhealth.exceptions.EntityNotFound;
+import br.com.dbc.wbhealth.model.dto.medico.MedicoAtendimentoDTO;
 import br.com.dbc.wbhealth.model.dto.medico.MedicoInputDTO;
 import br.com.dbc.wbhealth.model.dto.medico.MedicoOutputDTO;
 import br.com.dbc.wbhealth.model.entity.MedicoEntity;
@@ -11,6 +12,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -19,6 +21,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.time.LocalDate;
 import java.util.List;
 
 @Validated
@@ -63,5 +66,14 @@ public class MedicoController implements MedicoControllerDoc {
     public ResponseEntity<Void> deleteById(@PathVariable @Positive Integer idMedico) throws EntityNotFound {
         medicoService.delete(idMedico);
         return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/{medicoId}/atendimentos")
+    public ResponseEntity<List<MedicoAtendimentoDTO>> generateMedicoAtendimento(
+            @PathVariable Integer medicoId,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate dataInicio,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate dataFim) throws EntityNotFound {
+        List<MedicoAtendimentoDTO> report = medicoService.generateMedicoAtendimento(medicoId, dataInicio, dataFim);
+        return ResponseEntity.ok(report);
     }
 }
