+++ wbhealth/src/main/java/br/com/dbc/wbhealth/service/MedicoService.java
@@ -2,11 +2,14 @@ package br.com.dbc.wbhealth.service;
 
 import br.com.dbc.wbhealth.exceptions.EntityNotFound;
 import br.com.dbc.wbhealth.model.dto.hospital.HospitalOutputDTO;
+import br.com.dbc.wbhealth.model.dto.medico.MedicoAtendimentoDTO;
 import br.com.dbc.wbhealth.model.dto.medico.MedicoInputDTO;
 import br.com.dbc.wbhealth.model.dto.medico.MedicoOutputDTO;
+import br.com.dbc.wbhealth.model.dto.medico.PeriodoDTO;
 import br.com.dbc.wbhealth.model.entity.HospitalEntity;
 import br.com.dbc.wbhealth.model.entity.MedicoEntity;
 import br.com.dbc.wbhealth.model.entity.PessoaEntity;
+import br.com.dbc.wbhealth.repository.AtendimentoRepository;
 import br.com.dbc.wbhealth.repository.MedicoRepository;
 import br.com.dbc.wbhealth.repository.PessoaRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -17,6 +20,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -24,6 +29,7 @@ import java.util.List;
 public class MedicoService {
     private final MedicoRepository medicoRepository;
     private final PessoaRepository pessoaRepository;
+    private final AtendimentoRepository atendimentoRepository;
     private final HospitalService hospitalService;
     private final ObjectMapper objectMapper;
 
@@ -115,6 +121,28 @@ public class MedicoService {
         medicoOutput.setIdPessoa(pessoa.getIdPessoa());
 
         return medicoOutput;
+    }
+
+    public List<MedicoAtendimentoDTO> generateMedicoAtendimento(Integer idMedico, LocalDate dataInicio, LocalDate dataFim) throws EntityNotFound {
+        MedicoEntity medico = getMedicoById(idMedico);
+        if(medico == null) {
+            throw new EntityNotFound("Médico não encontrado com o ID: " + idMedico);
+        }
+
+        Long quantidadeAtendimentos = atendimentoRepository.countAtendimentosByMedicoAndDateRange(medico, dataInicio, dataFim);
+
+        List<MedicoAtendimentoDTO> atendimento = new ArrayList<>();
+        MedicoAtendimentoDTO medicoAtendimentoDTO = new MedicoAtendimentoDTO();
+        medicoAtendimentoDTO.setNomeMedico(medico.getPessoa().getNome());
+        medicoAtendimentoDTO.setCrm(medico.getCrm());
+        medicoAtendimentoDTO.setQuantidadeAtendimentos(quantidadeAtendimentos);
+
+        PeriodoDTO periodoDTO = new PeriodoDTO(dataInicio, dataFim);
+        medicoAtendimentoDTO.setPeriodo(periodoDTO);
+
+        atendimento.add(medicoAtendimentoDTO);
+
+        return atendimento;
     }
 
 }
